Use FadeOutLabel in TabViewCell
Extract FadeOutLabel to UIComponents
